+++ packages/webblocks-cms/src/Support/Plugins/Catalog/PluginCatalogDetailResult.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace WebBlocks\Cms\Support\Plugins\Catalog;
+
+class PluginCatalogDetailResult
+{
+  public function __construct(
+    public readonly bool $available,
+    public readonly ?CatalogPlugin $plugin,
+    public readonly string $baseUrl,
+    public readonly string $cmsVersion,
+    public readonly ?string $message = null,
+  ) {}
+}